Queue Manager FolderColor refactor Sprint 4 US 4
This refactor now pulls folder colors in Queue Manager from the SurveyProperty rules stored in QualPro_Params for the time being.  A combination of the new SurveyProperty function with an additional QMFolderColorByPrioriry function makes it possible to change the folder color that appears in QueueManager by Survey Type, or also by lateness criteria which may be customized per survey id.  Red icons were introduced to provide a way to visually flag surveys meeting the lateness criteria.  In addition, ICH survey types were given a FolderColor of orange as requested.