Temporary Queue Manager 3.4.0.0 S4 US4 TEMPORARY
This version became necessary as a fall-back option once we reached the last day of the sprint and did not have a resolution to the refactored version failing to move jobs into the print queue.  It has been installed in TEST and STAGING as of right now, but a new version of the refactored one with the resolution is being checked in on the Sprint branch, and that should be considered the permanent version likely to arrive in production rather than this one, which should only be parked in master on my machine.