PCMH Survey Properties (from subtype) S6 US6
This was a minor evolutionary step of the survey rules API which has now incorporated an optional "override" string parameter to be used in place of survey type.  For the moment, the only use of this override is the survey SubType, but others could be added. Additionally, not all existing survey rules were modified to make use of the survey SubType override, but rather these were only introduced as needed (where the defaults differed from the existing CGCAHPS defaults).