S9 US12 T1,2,3 FormLayout v3.1 Dynamic Cover Ltr
Have successfully achieved saving, loading, displaying, and editing a TextBoxName property for text boxes.  So tasks 1 and 2 are done for sure.  Task 3 may be done as well...but I haven't had a chance to check it out thoroughly yet.  Will do that next.  Will also test Click-Once deploy of FormLayout which is now in GIT for the first time.