S9 US12 T4,5 Artifacts Cover Letter operational
Thanks to the existence of the "GraphicsTab" define constant already existing in FormLayout, this went in much more quickly than was estimated.  A bonus here was that I added a label on the logo editor dialog so the name of the logo could be easily seen.  This is in preparation for task 12.6, which will be a drop down selector showing all names of text boxes and graphics on the page.