S9 US12 T6,7 Item Selector & Print Mockup removal
Now have an item selector modeled after the method picker in Visual Studio which shows the name of the currently selected item, or lets the user select from the drop down list in order to navigate to that item.  Also removed the Artifacts flagged pages from showing up in Print Mockup.