S10 US4 T1,2 Fix for blanks #2;coverletter dups
Discovered that after the save, the local text boxes still had the goofy text.  Fixed that.  Also got to a point where the cover letter name validation is good, so no duplicates are allowed for renames for new ones, and no names with [ or ] are permitted since those characters have special use for Artifacts page names on the tabs.