S10 US4 T1 Validation messages use TextBoxNames
Dave G identified messages in validation which were continuing to use the text box IDs but now it makes sense to use the text box names (labels) when they exist instead.  Related to that was a change to use the cover letter name rather than the cover letter ID in these same messages.