S15 US6 T1,2 Medicare Management analysis design
This script contains the logic which needs to be followed to prevent CensusForced from being automatically set to false for when the ProportionCalcPct drops below 75% in cases in which CensusForced was clearly set to true by a user.  Going forward, a record of who set CensusForced to true should be kept and it should also be checked to see if it was a user rather than the system, and similarly prevent CensusForced from being automatically set to false in that new case.