S19 INC0042007 Fix for FrequencyTool: Windows 7+
Richard had set up a couple of 64-bit VMs with Windows 2012 and Frequency Tool was kicking out an error based on hard-coded paths for the location of the Users (formerly Documents and Settings) folder and Documents (formerly My Documents).  This is only a proof of concept situation...may need to revisit this if we don't port this tool.