S22 US12 T? Found PCLGen code and modified already
Early in the Sprint I found what appeared to be the code which was already keeping one comment box together with a question. After learning the ropes for how to unit test via the corresponding table, I was able to piece in the code to look for one more prior comment box to keep together with the original two being kept together. Dave G should review this code when he returns.