Fix: Sampling Service Octopus added settings.xml
There was no settings.xml being deployed with the service which disallows successful start of the service. I found that the same settings.xml from production was set up in Stage and Test on the prior version running there, and even though stage and test are now unable to access the production URL contained, the service launches, which is what we need for Octopus deploy to succeed.