RTP-2431 removed part of the WHERE clause when populating #Work
from Dana:
The original intent there was that we not update something to nothing. That was in place before the proc updated the HAadmissionSource and HDischargeStatus fields.

I had a chat with Rachel, and we both think we should just get rid of that. We should assume the update file has the correct data. We can trace the value in the database back to the file and it will make sense. If we ever get a file with a problem, like all the DRGs are missing, that will be apparent when the submission is QA’d, and they can work with the client to fix the problem.

So, let’s just comment out that where clause.